Refactor ActivityView
